+++ pumpkin-protocol/src/client/play/chunk_data.rs
@@ -5,7 +5,7 @@ use pumpkin_data::packet::clientbound::PLAY_LEVEL_CHUNK_WITH_LIGHT;
 use pumpkin_macros::packet;
 use pumpkin_world::{
     DIRECT_PALETTE_BITS,
-    chunk::{ChunkData, SUBCHUNKS_COUNT},
+    storage::{ChunkData, SUBCHUNKS_COUNT},
 };
 
 #[packet(PLAY_LEVEL_CHUNK_WITH_LIGHT)]
